Fix tests and docker connection
