+++ Dockerfile
@@ -13,4 +13,8 @@ CMD ["java","-jar","-Dspring.profiles.active=h2,cache,redis-cache","machines.jar
 
 
 # build docker: docker build -t machines:1.0.0 .
-# docker build -t
+
+# docker build -t habibahmedmagdy/machines:1.0.0 .
+# docker tag machines:2.0.0 habibahmedmagdy/machines:2.0.0
+# By using docker commit <existing-container> <hub-user>/<repo-name>[:<tag>] to commit changes
+# docker push habibahmedmagdy/machines:1.0.0
